dvfs: Change signature 'struct inode_operations::trancate'

+++ src/compat/posix/sys/mman/dvfs_mman.c
@@ -1,3 +1,5 @@
+#include <sys/types.h>
+
 #include <mem/vmem.h>
 #include <mem/phymem.h>
 #include <util/binalign.h>
@@ -11,7 +13,7 @@ size_t shm_get_len(struct shm *shm) {
 	return shm->inode.length;
 }
 
-int shm_truncate(struct inode *inode, size_t len) {
+int shm_truncate(struct inode *inode, off_t len) {
 	void *phy;
 	len = binalign_bound(len, VMEM_PAGE_SIZE);
