+++ src/fs/driver/fat/fat_dvfs.c
@@ -13,6 +13,7 @@
 #include <stdlib.h>
 #include <fcntl.h>
 #include <limits.h>
+#include <sys/types.h>
 
 #include <drivers/device.h>
 
@@ -136,7 +137,7 @@ static int fat_create(struct inode *i_new, struct inode *i_dir, int mode) {
 	return fat_create_file(fi, di, name, mode);
 }
 
-static int fat_truncate(struct inode *inode, size_t len) {
+static int fat_truncate(struct inode *inode, off_t len) {
 	/* This is a stub, but files should be extended automatically
 	 * with the common part of the driver on write */
 	return 0;
